+++ packages/moonstone/Scroller/Scrollable.js
@@ -631,7 +631,8 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 					canScrollVertically = this.canScrollVertically(bounds),
 					delta = this.wheel(e, canScrollHorizontally, canScrollVertically),
 					direction = Math.sign(delta),
-					focusedItem = Spotlight.getCurrent();
+					focusedItem = Spotlight.getCurrent(),
+					scrollBarRef = this.verticalScrollbarRef || this.horizontalScrollbarRef;
 
 				Spotlight.setPointerMode(false);
 				if (focusedItem) {
@@ -639,6 +640,9 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 				}
 
 				this.childRef.setContainerDisabled(true);
+				if (scrollBarRef) {
+					scrollBarRef.setWheelMode(true);
+				}
 
 				if (direction !== this.wheelDirection) {
 					this.isScrollAnimationTargetAccumulated = false;
@@ -788,6 +792,8 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 		}
 
 		stop () {
+			const scrollBarRef = this.verticalScrollbarRef || this.horizontalScrollbarRef;
+
 			this.animator.stop();
 			this.isScrollAnimationTargetAccumulated = false;
 			this.childRef.setContainerDisabled(false);
@@ -796,6 +802,10 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 			this.lastScrollPositionOnFocus = null;
 			this.hideThumb();
 			this.doScrollStop();
+
+			if (scrollBarRef) {
+				scrollBarRef.setWheelMode(false);
+			}
 		}
 
 		focusOnItem () {